done with all the history stuff

+++ app/src/main/java/com/example/foodapp2025/data/model/CartModel.java
@@ -7,13 +7,13 @@ import java.util.Arrays;
 public class CartModel {
     private String imageUrl;
     private String name;
-    private int price;
-    private int quantity;
+    private Long price;
+    private Long quantity;
     private static final double TAX_RATE = 0.1; // 10 % tax
     private static final Integer DELIVERY_FEE = 20000; // Fixed delivery fee
     private static final ArrayList<String> VOUCHER_CODE = new ArrayList<>(Arrays.asList("FREE2SHIP", "10DEAL"));
 
-    public CartModel(String imageUrl, String name, int price, int quantity) {
+    public CartModel(String imageUrl, String name, Long price, Long quantity) {
         this.imageUrl = imageUrl;
         this.name = name;
         this.price = price;
@@ -28,15 +28,15 @@ public class CartModel {
         return imageUrl;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public int getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
